Added better handler for File Deletion
*	Added support for generating `OtherImages` (Those that are not Visible, Infared, WaterVapour or False Color)

For the scenario where the False Color is disabled, after generation of each part
(Visible, Infrared, WaterVapour, OtherImages) the ImageManager will delete the files
used for the generation of the images. In the scenario where False Color is enabled,
It will wait until the False Color is generated to erase each part. In the case of images
that will not have either Visible or Infrared (so no posibility of generating False Color),
there will be a timeout of 2h to erase the files. If in 2 hours it doesn't generate the
required images, if EraseFiles is enabled, it will erase the corresponding .lrit files.

This closes #27